refactor(kv): consolidate backend interface
